Hiephd dev

+++ services/submission-service/src/main/java/com/submission_service/repository/IFeedBackAiRepository.java
@@ -3,5 +3,5 @@ package com.submission_service.repository;
 import com.submission_service.model.entity.FeedbackAi;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IFeedBackAiRepository extends JpaRepository<FeedbackAi, Integer> {
+public interface IFeedBackAiRepository extends JpaRepository<FeedbackAi, Long> {
 }
